move initialize using query parameters convention to rich transient feature
